Switched out some images for Ferris the crab.
On call with a friend, they mentioned this extension to me and gave me a great idea for how to learn a bit of TypeScript.
To be clear, the idea to replace change the extension to use Ferris the crab instead of Mr. Incredible was entirely theirs, not mine.
I am not mentioning their name purely for their privacy.

I'm not sure how much work I'll do on this, if any, but it was a great distraction!